updated student file submision with dropzone

+++ client/src/components/StudentGroup/Submissions.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import React, { useState, useEffect, useCallback } from "react";
 import "./Submissions.css";
 import Box from "@mui/material/Box";
 import { Chip, Divider } from "@mui/material";
@@ -26,8 +26,42 @@ import IconButton from "@mui/material/IconButton";
 import CloseIcon from "@mui/icons-material/Close";
 import { triggerBase64Download } from "common-base64-downloader-react";
 import axios from "axios";
-
+import Dropzone from "react-dropzone";
 const API = process.env.REACT_APP_API;
+import React from "react";
+import { useDropzone } from "react-dropzone";
+import styledComponents from "styled-components";
+
+//colors for dropzone
+const getColor = (props) => {
+  if (props.isDragAccept) {
+    return "#00e676";
+  }
+  if (props.isDragReject) {
+    return "#ff1744";
+  }
+  if (props.isFocused) {
+    return "#2196f3";
+  }
+  return "#eeeeee";
+};
+
+//styles for dropzone
+const Container = styledComponents.div`
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 20px;
+  border-width: 2px;
+  border-radius: 2px;
+  border-color: ${(props) => getColor(props)};
+  border-style: dashed;
+  background-color: #fafafa;
+  color: #bdbdbd;
+  outline: none;
+  transition: border 0.24s ease-in-out;
+`;
 
 export default function Submissions({ studentGroup }) {
   const [Group, setGroup] = useState(studentGroup);
@@ -38,9 +72,59 @@ export default function Submissions({ studentGroup }) {
     useState(false);
   const [studentSubmission, setStudentSubmission] = useState({});
   const [existingSubmission, setExistingSubmission] = useState(null);
+  const [existingSubmissionName, setExistingSubmissionName] = useState(null);
+
   const [base64File, setBase64File] = useState();
   const [fileIsLoadig, setFileIsLoading] = useState(true);
   const [viewFeedbackModal, setviewFeedbackModal] = useState(false);
+
+  //new styled dropzone component
+  function StyledDropzone(props) {
+    let submissionObject = {
+      name: "",
+      base64: "",
+    };
+
+    //callback function to handle dropzone inputs
+    const onDrop = useCallback((acceptedFiles) => {
+      const file = acceptedFiles[0];
+      submissionObject.name = file.name;
+      console.log(submissionObject.name);
+
+      const reader = new FileReader();
+      reader.onabort = () => console.log("file reading was aborted");
+      reader.onerror = () => console.log("file reading has failed");
+      reader.onload = () => {
+        // Do whatever you want with the file contents
+        submissionObject.base64 = reader.result;
+        console.log(submissionObject.base64);
+      };
+      reader.readAsDataURL(file);
+      setStudentSubmission(submissionObject);
+      setExistingSubmissionName(submissionObject.name);
+    }, []);
+    const {
+      getRootProps,
+      getInputProps,
+      isFocused,
+      isDragAccept,
+      isDragReject,
+    } = useDropzone({
+      // accept: {
+      //   "application/pdf": [],
+      // },
+      onDrop,
+    });
+
+    return (
+      <div className="container">
+        <Container {...getRootProps({ isFocused, isDragAccept, isDragReject })}>
+          <input {...getInputProps()} />
+          <p>Drag and drop some files here, or click to select files</p>
+        </Container>
+      </div>
+    );
+  }
 
   // Submission Types
   // TopicAssesmentForm
@@ -183,7 +267,10 @@ export default function Submissions({ studentGroup }) {
         submissionObj
       );
       setExistingSubmission(result.data.data);
-      setBase64File(result.data.data.file.base64);
+      if (result.data.data) {
+        setBase64File(result.data.data.file.base64);
+        setExistingSubmissionName(result.data.data.file.name);
+      }
       // console.log("olaaaaaaa" + result.data.data);
       // setPageIsLoading(false);
     } catch (err) {
@@ -234,19 +321,21 @@ export default function Submissions({ studentGroup }) {
   //function to upload file
   const onFileUpload = (file) => {
     setStudentSubmission(file);
-    console.log(file);
   };
 
   //function to delete student submission from db
   const deleteStudentSubmission = async () => {
     setFileIsLoading(true);
-    const id = existingSubmission._id;
-    const result = await axios.delete(`${API}/studentSubmission/${id}`);
+    if (existingSubmission) {
+      const id = existingSubmission._id;
+      const result = await axios.delete(`${API}/studentSubmission/${id}`);
+    }
     setFileIsLoading(false);
     setExistingSubmission(null);
+    setExistingSubmissionName(null);
   };
 
-  //function to download uploaded student submissions
+  // // function to download uploaded student submissions
   // const downloadStudentSubmission = async () => {
   //   const base64File = existingSubmission.file.base64;
   // };
@@ -403,7 +492,7 @@ export default function Submissions({ studentGroup }) {
                   <CircularProgress color="inherit" />
                 ) : (
                   <div>
-                    {existingSubmission ? (
+                    {existingSubmissionName ? (
                       <>
                         <Button
                           onClick={() => {
@@ -420,18 +509,20 @@ export default function Submissions({ studentGroup }) {
                           onClick={() => {
                             triggerBase64Download(
                               base64File,
-                              existingSubmission.file.name
+                              existingSubmissionName
                             );
                           }}
                         >
-                          {existingSubmission.file.name}
+                          {existingSubmissionName}
                         </Button>
                       </>
                     ) : (
-                      <FileBase64
-                        multiple={false}
-                        onDone={onFileUpload.bind()}
-                      />
+                      // <FileBase64
+                      //   multiple={false}
+                      //   onDone={onFileUpload.bind()}
+                      // />
+
+                      <StyledDropzone />
                     )}
                   </div>
                 )}
